Less casting on numbers

+++ LifeKit/Sources/LifeKit/LifeController.swift
@@ -9,22 +9,22 @@
 import Foundation
 
 public final class LifeController: NSObject {
-    private static let gridSize: NSInteger = 15
+    private static let gridSize: CGFloat = 15.0
     
     var grid: Grid
     var gridView: GridView
     
     public init(gridView: GridView) {
-        self.grid = Grid(sideLength: Self.gridSize)
+        self.grid = Grid(sideLength: Int(Self.gridSize))
             .fill(with: AbstractCell.deadCell)
         self.gridView = gridView
         
         _ = self.gridView.drawGrid(grid)
     }
     
-    func changeCellAtRelative(x fractionX: Float, y fractionY: Float) -> Self {
-        let cellX = Int(fractionX * Float(Self.gridSize))
-        let cellY = Int(fractionY * Float(Self.gridSize))
+    func changeCellAtRelative(x fractionX: CGFloat, y fractionY: CGFloat) -> Self {
+        let cellX = Int(fractionX * Self.gridSize)
+        let cellY = Int(fractionY * Self.gridSize)
         
         grid = grid.at(x: cellX, y: cellY, perform: {
             $0?.changePopulation()
